Refactor updateOrderTaxTxn function
