+++ order-syncer/src/clients/update.client.js
@@ -3,7 +3,14 @@ import CustomError from '../errors/custom.error.js';
 import { HTTP_STATUS_SUCCESS_ACCEPTED } from '../constants/http.status.constants.js';
 import { getOrder } from './query.client.js';
 
-export async function updateOrder(actions, orderId) {
+export async function updateOrderTaxTxn(taxTxnId, orderId) {
+  let actions = [];
+  actions.push({
+    action: 'setCustomField',
+    name: 'taxTransactionReference',
+    value: taxTxnId,
+  });
+
   const order = await getOrder(orderId);
   return await createApiRoot()
     .orders()
